Move MappingVirtualProperty document classes to the PropertyHooksTest to prevent invalid document from polluting the other tests
