+++ tests/Tests/Functional/PropertyHooksTest.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace Doctrine\ODM\MongoDB\Tests\Functional;
 
+use Doctrine\ODM\MongoDB\Mapping\Annotations\Document;
+use Doctrine\ODM\MongoDB\Mapping\Annotations\Field;
+use Doctrine\ODM\MongoDB\Mapping\Annotations\Id;
 use Doctrine\ODM\MongoDB\Mapping\MappingException;
 use Doctrine\ODM\MongoDB\Tests\BaseTestCase;
-use Documents\PropertyHooks\MappingVirtualProperty;
 use Documents\PropertyHooks\User;
 
 class PropertyHooksTest extends BaseTestCase
@@ -91,8 +93,29 @@ class PropertyHooksTest extends BaseTestCase
     public function testMappingVirtualPropertyIsNotSupported(): void
     {
         $this->expectException(MappingException::class);
-        $this->expectExceptionMessage('Mapping virtual property "fullName" on document "Documents84\PropertyHooks\MappingVirtualProperty" is not allowed.');
+        $this->expectExceptionMessage('Mapping virtual property "fullName" on document "' . __NAMESPACE__ . '\MappingVirtualProperty" is not allowed.');
 
         $this->dm->getClassMetadata(MappingVirtualProperty::class);
+    }
+}
+
+#[Document(collection: 'property_hooks_user')]
+class MappingVirtualProperty
+{
+    #[Id]
+    public ?string $id;
+
+    #[Field]
+    public string $first;
+
+    #[Field]
+    public string $last;
+
+    #[Field]
+    public string $fullName {
+        get => $this->first . " " . $this->last;
+        set {
+            [$this->first, $this->last] = explode(' ', $value, 2);
+        }
     }
 }
